requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.47.0
pyyaml>=3.13
aiohttp>=3.1.3
python-socketio==4.6.1
pyserial
